minor refactor: rename \<MDXComponents => useMDXComponents
